remove mainnet2, some script modifications
