--- rust/sealevel/environments/mainnet2/.gitignore
@@ -1,1 +1,0 @@
-**/**/keys
