--- rust/sealevel/environments/mainnet2/gas-oracle-configs.json
@@ -1,20 +1,0 @@
-[
-    {
-        "domain": 56,
-        "gasOracle": {
-            "type": "remoteGasData",
-            "tokenExchangeRate": "100000000000000000000",
-            "gasPrice": "3000000000",
-            "tokenDecimals": 18
-        }
-    },
-    {
-        "domain": 22222,
-        "gasOracle": {
-            "type": "remoteGasData",
-            "tokenExchangeRate": "4700000000000000",
-            "gasPrice": "1000000000",
-            "tokenDecimals": 18
-        }
-    }
-]
