--- rust/sealevel/environments/mainnet2/multisig-ism-message-id/solana/nautilus/program-ids.json
@@ -1,3 +1,0 @@
-{
-  "program_id": "9k74DkJvS2x9QhG4XfnKsLkqaCDyVfaj8s6FyJyhAeEP"
-}
