--- rust/sealevel/environments/mainnet2/multisig-ism-message-id/solana/rc/program-ids.json
@@ -1,3 +1,0 @@
-{
-  "program_id": "BYTsxBuKVbwgsZFswzB91nrxveQySghwXzaKqn8exNnC"
-}
